Vocabulary view reworked
The word list is now more compact, and there are two view modes for the vocabulary view. List view is a list of entries, where you can also delete words and have access to information like when a word was added. Learn view is just a table with source word and translated word, where the translation is covered until the user touches to remove the cover. This is useful for learning the words and trying to translate them yourself.